+++ CHANGELOG.md
@@ -7,13 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## TO-DOs
 
-- Add analyzer tests.
 - Add frontend tests.
 
 ## [Unreleased] XX-XX-XXXX
 
 ### Added
 
+- Analyzer tests
 - Backend tests.
 - Github actions: Pylint.